Moved Sets, Update, and MapTools classes to utils subpackage.
