+++ src/main/java/net/christophschubert/kafka/clusterstate/ClusterStateDiff.java
@@ -1,8 +1,10 @@
 package net.christophschubert.kafka.clusterstate;
 
 
-import net.christophschubert.kafka.clusterstate.mds.RbacBinding;
 import net.christophschubert.kafka.clusterstate.mds.RbacBindingInScope;
+import net.christophschubert.kafka.clusterstate.utils.MapTools;
+import net.christophschubert.kafka.clusterstate.utils.Sets;
+import net.christophschubert.kafka.clusterstate.utils.Update;
 
 import java.util.HashMap;
 import java.util.Map;
