+++ src/main/java/net/christophschubert/kafka/clusterstate/policies/ConfigUpdateDoNothingPolicy.java
@@ -1,6 +1,6 @@
 package net.christophschubert.kafka.clusterstate.policies;
 
-import net.christophschubert.kafka.clusterstate.Update;
+import net.christophschubert.kafka.clusterstate.utils.Update;
 import net.christophschubert.kafka.clusterstate.actions.Action;
 
 import java.util.Collections;
