+++ src/main/java/net/christophschubert/kafka/clusterstate/utils/MapTools.java
@@ -1,4 +1,4 @@
-package net.christophschubert.kafka.clusterstate;
+package net.christophschubert.kafka.clusterstate.utils;
 
 
 import org.slf4j.Logger;
